Tweaked the backup/restore scripts

+++ roles/thepoint_website/files/upperroom-restore.sh
@@ -6,17 +6,17 @@ DB_CONTAINER="${1?No DB container specified}"
 APP_CONTAINER="${2?No app container specified}"
 MEDIA_BUCKET="${3?No media bucket specified}"
 BACKUP_BUCKET="${4?No backup bucket specified}"
-BACKUP_FILE_DB="${5-thepoint.pgsql}"
-BACKUP_FILE_MEDIA="${6-media}"
-CACHE_CONTAINER="${7}"
+CACHE_CONTAINER="${5}"
+BACKUP_FILE_DB="${BACKUP_FILE_DB-thepoint.pgsql}"
+BACKUP_FILE_MEDIA="${BACKUP_FILE_MEDIA-media}"
+DB_NAME="${DB_NAME-thepoint}"
+DB_USER="${DB_USER-thepoint}"
 
 docker stop "${APP_CONTAINER}"
-aws s3 sync "s3://${BACKUP_BUCKET}/${BACKUP_FILE_MEDIA}/" "s3://${MEDIA_BUCKET}/" --delete
-docker exec -i "${DB_CONTAINER}" psql -U thepoint -d postgres -c 'DROP DATABASE thepoint'
-docker exec -i "${DB_CONTAINER}" psql -U thepoint -d postgres -c 'CREATE DATABASE thepoint'
-aws s3 cp "s3://${BACKUP_BUCKET}/${BACKUP_FILE_DB}" - | docker exec -i "${DB_CONTAINER}" pg_restore --no-acl --no-owner -U thepoint -d thepoint
-if [ -n "${CACHE_CONTAINER}" ]; then
-    docker exec -t "${CACHE_CONTAINER}" redis-cli FLUSHALL
-fi
+[ -z "${DB_ONLY}" ] && aws s3 sync "s3://${BACKUP_BUCKET}/${BACKUP_FILE_MEDIA}/" "s3://${MEDIA_BUCKET}/" --delete
+docker exec -i "${DB_CONTAINER}" psql -U "${DB_USER}" -d postgres -c "DROP DATABASE ${DB_NAME}"
+docker exec -i "${DB_CONTAINER}" psql -U "${DB_USER}" -d postgres -c "CREATE DATABASE ${DB_NAME}"
+aws s3 cp "s3://${BACKUP_BUCKET}/${BACKUP_FILE_DB}" - | docker exec -i "${DB_CONTAINER}" pg_restore --no-acl --no-owner -U "${DB_USER}" -d "${DB_NAME}"
+[ -n "${CACHE_CONTAINER}" ] && docker exec -t "${CACHE_CONTAINER}" redis-cli FLUSHALL
 docker start "${APP_CONTAINER}"
 docker exec "${APP_CONTAINER}" upperroom migrate --no-input
